04_AutoMLLink/formula/FormulaAdvisor/requirements.txt
autokeras==1.0.20
category_encoders==2.6.1
mysql_connector_python==8.0.33
numpy==1.23.5
pandas==2.0.1
redis==4.5.5
requests==2.30.0
scikit_learn==1.2.2
scipy==1.10.1
# tensorflow==2.12.0
